DB상 speed Double화 / 평속 계산시 첫 속도는 무조건 들어가게

+++ app/src/main/java/kr/musekee/faremeter/libs/DBDataClasses.kt
@@ -13,6 +13,6 @@ data class DrivingData(
     val id: String,
     val latitude: Double, // 37.00001
     val longitude: Double, // 130.75745
-    val speed: Float, // m/s
+    val speed: Double, // m/s
     val time: Long // ms
 )
